chore: reorganize ports
